pyproject.toml
[tool.poetry]
name = "ureport"
version = "1.2.51+patch1"
description = "U-Report dashboard built on data collected by RapidPro"
authors = ["Nyaruka Ltd <code@nyaruka.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.8"
smartmin = "^4.0.0"
rapidpro-dash = "^1.9.3"
colorama = "^0.4.3"
django-compressor = "^3.0"
django-debug-toolbar = "3.2.4"
django-digest = "^1.13"
django-rosetta = "^0.8.1"
django-storages = "^1.6.6"
django-rest-swagger = "^2.2.0"
djangorestframework = "3.11.2"
dj-database-url = "^0.5.0"
boto3 = "^1.7.67"
Pillow = "^9.1.1"
gunicorn = "^20.1.0"
pisa = "^3.0.33"
sentry-sdk = "^0.19.5"
django-redis = "^4.11.0"
psycopg2-binary = "^2.8.6"
xlutils = "^2.0.0"
mock = "^2.0.0"
django-hamlpy = "^1.2"
pycountry = "^18.5.26"
django-countries = "^7.2.1"
GitPython = "^2.1.11"
regex = "^2020.9.27"
rapidpro-python = "^2.8.5"
feedparser = "^6.0.2"
Markdown = "^2.6.11"
iso8601 = "^0.1.12"
sorl-thumbnail = "^12.7.0"
stop_words = "^2018.7.23"
celery = {extras = ["redis"], version = ">4.4.6"}


[tool.poetry.dev-dependencies]
black = "^22.3"
isort = "^5.9.3"
flake8 = "^3.7.9"
codecov = "^2.1.11"

[tool.black]
line-length = 119

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
